+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.EXOActiveSyncDeviceAccessRule.Tests.ps1
@@ -39,6 +39,8 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             # Mock Write-Host to hide output during the tests
             Mock -CommandName Write-Host -MockWith {
             }
+            $Script:exportedInstances =$null
+            $Script:ExportMode = $false
         }
 
         # Test contexts
@@ -54,12 +56,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 }
 
                 Mock -CommandName Get-ActiveSyncDeviceAccessRule -MockWith {
-                    return @{
-                        Identity       = 'iOS 6.1 10B145 (WrongOS)'
-                        AccessLevel    = 'Allow'
-                        Characteristic = 'DeviceOS'
-                        QueryString    = 'iOS 6.1 10B145'
-                    }
+                    return $null
                 }
                 Mock -CommandName Set-ActiveSyncDeviceAccessRule -MockWith {
                     return @{