+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.EXOMailboxSettings.Tests.ps1
@@ -34,6 +34,8 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             # Mock Write-Host to hide output during the tests
             Mock -CommandName Write-Host -MockWith {
             }
+            $Script:exportedInstances =$null
+            $Script:ExportMode = $false
         }
 
         Context -Name 'Specified TimeZone is Invalid' -Fixture {