+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.PlannerTask.Tests.ps1
@@ -38,6 +38,8 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             # Mock Write-Host to hide output during the tests
             Mock -CommandName Write-Host -MockWith {
             }
+            $Script:exportedInstances =$null
+            $Script:ExportMode = $false
 
             Mock -CommandName Get-MgUser -MockWith {
                 return @{